refactor(challenge): tooltip을 공용 폴더 대신 사용처로 이동 (usage-based 배치)
ui/tooltip-question/ 묶음 폴더를 없애고 실사용처로:
- ScoreTooltipQuestion → dashboard/ui/ (dashboard/section/ScoreSection만 사용)
- MissionTooltipQuestion → ui/ (challenge 페이지에서 사용)
import 경로 갱신. tsc 0 / lint 통과.
