+++ apps/web/src/domain/challenge/dashboard/ui/ScoreTooltipQuestion.tsx
