+++ apps/web/src/domain/challenge/my-challenge/section/mission-submit/MissionSubmitBonusSection.tsx
@@ -2,10 +2,10 @@ import { twMerge } from '@/lib/twMerge';
 import { Schedule } from '@/schema';
 import { clsx } from 'clsx';
 import { ReactNode } from 'react';
-import LinkChangeConfirmationModal from '../../../modal/LinkChangeConfirmationModal';
+import LinkChangeConfirmationModal from '../../../ui/LinkChangeConfirmationModal';
 import BankSelectDropdown from '../../mission/BankSelectDropdown';
-import MissionSubmitButton from '../../mission/ui/MissionSubmitButton';
 import MissionToast from '../../mission/MissionToast';
+import MissionSubmitButton from '../../mission/ui/MissionSubmitButton';
 import LinkInputSection from '../LinkInputSection';
 import AgreementSection from '../mission-submit-bonus/AgreementSection';
 import { useMissionSubmitBonus } from '../mission-submit-bonus/useMissionSubmitBonus';
