+++ apps/web/src/domain/challenge/ui/MissionTooltipQuestion.tsx
